linux-yocto/tiny: switch default branch to standard/tiny/base
The default KBRANCH for building -tiny should not be machine specific.
In fact, the SRCREVs that have been in place for -tiny have been using
the hashes from tiny/base for some time, but since the machine specific
branches all contain that hash, the mismatch hasn't been spotted.

Setting the default branch to standard/tiny/base makes sure that the
default branch and SRCREVs always match.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>